Remove versioned subscription support
