--- src/NServiceBus.RavenDB/Subscriptions/ISubscriptionIdFormatter.cs
@@ -1,9 +1,0 @@
-﻿namespace NServiceBus.RavenDB.Persistence.SubscriptionStorage
-{
-    using NServiceBus.Unicast.Subscriptions;
-
-    internal interface ISubscriptionIdFormatter
-    {
-        string FormatId(MessageType messageType);
-    }
-}
