+++ src/NServiceBus.RavenDB/Subscriptions/SubscriptionPersister.cs
@@ -3,6 +3,8 @@ namespace NServiceBus.Persistence.RavenDB
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Security.Cryptography;
+    using System.Text;
     using System.Threading.Tasks;
     using System.Transactions;
     using NServiceBus.Extensibility;
@@ -18,12 +20,10 @@ namespace NServiceBus.Persistence.RavenDB
         public SubscriptionPersister(IDocumentStore store)
         {
             documentStore = store;
-            SubscriptionIdFormatter = new VersionedSubscriptionIdFormatter();
         }
 
         public TimeSpan AggressiveCacheDuration { get; set; } = TimeSpan.FromMinutes(1);
         public bool DisableAggressiveCaching { get; set; }
-        public ISubscriptionIdFormatter SubscriptionIdFormatter { get; set; }
 
         public async Task Subscribe(Subscriber subscriber, MessageType messageType, ContextBag context)
         {
@@ -41,7 +41,7 @@ namespace NServiceBus.Persistence.RavenDB
                 {
                     using (var session = OpenAsyncSession())
                     {
-                        var subscriptionDocId = SubscriptionIdFormatter.FormatId(messageType);
+                        var subscriptionDocId = GetDocumentIdForMessageType(messageType);
 
                         var subscription = await session.LoadAsync<Subscription>(subscriptionDocId).ConfigureAwait(false);
 
@@ -88,7 +88,7 @@ namespace NServiceBus.Persistence.RavenDB
 
             using (var session = OpenAsyncSession())
             {
-                var subscriptionDocId = SubscriptionIdFormatter.FormatId(messageType);
+                var subscriptionDocId = GetDocumentIdForMessageType(messageType);
 
                 var subscription = await session.LoadAsync<Subscription>(subscriptionDocId).ConfigureAwait(false);
 
@@ -108,7 +108,7 @@ namespace NServiceBus.Persistence.RavenDB
 
         public async Task<IEnumerable<Subscriber>> GetSubscriberAddressesForMessage(IEnumerable<MessageType> messageTypes, ContextBag context)
         {
-            var ids = messageTypes.Select(messageType => SubscriptionIdFormatter.FormatId(messageType)).ToList();
+            var ids = messageTypes.Select(GetDocumentIdForMessageType).ToList();
 
             using (var suppressTransaction = new TransactionScope(TransactionScopeOption.Suppress, TransactionScopeAsyncFlowOption.Enabled))
             {
@@ -129,6 +129,27 @@ namespace NServiceBus.Persistence.RavenDB
 
                 suppressTransaction.Complete();
                 return subscribers;
+            }
+        }
+
+        string GetDocumentIdForMessageType(MessageType messageType)
+        {
+            using (var provider = new SHA1CryptoServiceProvider())
+            {
+                var inputBytes = Encoding.UTF8.GetBytes(messageType.TypeName);
+                var hashBytes = provider.ComputeHash(inputBytes);
+
+                // 54ch for perf - "Subscriptions/" (14ch) + 40ch hash
+                var idBuilder = new StringBuilder(54);
+
+                idBuilder.Append("Subscriptions/");
+
+                for (var i = 0; i < hashBytes.Length; i++)
+                {
+                    idBuilder.Append(hashBytes[i].ToString("x2"));
+                }
+
+                return idBuilder.ToString();
             }
         }
